y4m: fix parser error for y4m input
EB_STRNCPY assume the dest is a static array, so it will use sizeof(dst) to get the size
But it's not ture in copyUntilCharacterOrNewLine. So the y4m as input is totally breaking.

Since we can't check the type of dst, assume it's static array is harmful.
Better let user pass the dst size.